Reformat commands (within Cluster) test codes
